Basic implementation of fizzbuzz, with only limited options.
In particular, there's currently no way to specify which numbers get
substituted out, what the words are that get substituted in for them,
or which colors are used and in what order: the program just uses its
built-in harcoded defaults (multiples of 3 become Fizz, multipes of 5
become Buzz, and the standard ANSI colors) no matter what.  Also,
there's no support for languages other than English as yet.

The only options you can change are what number to count from, what
number to count to, the separator, whether to spell out the numbers,
whether to use hyphens, and whether to do word wrap.

So it's pretty basic, but it works and technically meets the
requirements, so it's a start.